+++ resources/webxr.js
@@ -9,6 +9,31 @@ window.addEventListener('enter-vr', e => {
 });
 
 // Components
+AFRAME.registerComponent('toggle-handler', {
+  schema: {
+    target: { default: '#' }
+  },
+  init: function () {
+    var data = this.data;
+    //alert(data.target);
+    var toggle_obj = document.getElementById(data.target);
+    if (toggle_obj) {
+      toggle_obj.setAttribute("visible", "false");
+    }
+
+    this.el.addEventListener('click', function () {
+      var toggle_obj = document.getElementById(data.target);
+      let value = toggle_obj.getAttribute("visible");
+      if (value == true) {
+        toggle_obj.setAttribute("visible", "false");
+      }
+      else {
+        toggle_obj.setAttribute("visible", "true");
+      }
+    })
+  }
+});
+
 AFRAME.registerComponent('link-handler', {
   schema: {
     target: { default: '#' }
@@ -65,35 +90,35 @@ AFRAME.registerComponent('light-map-geometry', {
   schema: {
     path: { default: '' },
     format: { default: 'RGBFormat' },
-	intensity: { default: 1.0}
+    intensity: { default: 1.0 }
   },
 
   init: function () {
     const data = this.data;
     const el = this.el;
     this.texture = new THREE.TextureLoader().load(data.path);
-	this.intensity = data.intensity;
-    this.applyLightMap();	
+    this.intensity = data.intensity;
+    this.applyLightMap();
     this.el.addEventListener('object3dset', this.applyLightMap.bind(this));
   },
 
   applyLightMap: function () {
     const mesh = this.el.getObject3D('mesh');
     const lightMap = this.texture;
-	this.texture.flipY = false;
+    this.texture.flipY = false;
     const el = this.el
-	const value = this.intensity;
+    const value = this.intensity;
 
     if (!mesh) return;
     mesh.traverse(function (node) {
       //if (node.geometry && node.geometry.type == "BufferGeometry") {
-        //console.log(node);
-        //console.log(node.geometry.attributes);
-        //node.geometry.attributes.uv2 = node.geometry.attributes.uv.clone();
+      //console.log(node);
+      //console.log(node.geometry.attributes);
+      //node.geometry.attributes.uv2 = node.geometry.attributes.uv.clone();
       //}
       if (node.material && 'lightMap' in node.material) {
         node.material.lightMap = lightMap;
-		node.material.lightMapIntensity = value;
+        node.material.lightMapIntensity = value;
         node.material.needsUpdate = true;
       }
     });